ME_QEATS_V2_MODULE_REDIS

+++ qeatsbackend/src/main/java/com/crio/qeats/configs/RedisConfiguration.java
@@ -79,8 +79,10 @@ public class RedisConfiguration {
   public void destroyCache() {
 
     if (this.jedisPool != null) {
-      jedisPool.getResource().flushAll();
-      jedisPool.close();
+      if (jedisPool.getResource() != null) {
+        jedisPool.getResource().flushAll();
+      }
+      jedisPool.destroy();
       jedisPool = null;
     }
